+++ FibonacciSequence/Interfaces/IFibonacciSequencer.cs
@@ -2,8 +2,8 @@
 {
     public interface IFibonacciSequencer
     {
-        int NextElement(int n_1, int n_2);
-        int[] GenerateSequence(int n);
+        ulong NextElement(ulong n_1, ulong n_2);
+        ulong[] GenerateSequence(int n);
         string GetSequenceString(int n, char separator);
     }
 }
